Testing for York fitting implemented as part of the OCO package
The unit tests in oco_validation/tests/test_fitting.py currently
run the York fitting from jllutils.stats with both artificial data
with known slopes and real example TCCON/OCO-2 data. For the artificial
data, it checks that the returned slopes and intercepts are what
they were set to when creating the data. For all data, it also checks
that it returns the same slope, intercept, and errors as the Matlab
function in D. Wunch's original code.

Two other fixes:
    * Corrected a circular import between data.containers, validation.__init__,
and validation.containers, which happened because data.containers was importing
InsufficientTCCONDataError from validation, which required importing __init__,
which imported validation.containers.TargetMatch, which imported data.containers.
Contents of validation.__init__ moved to validation.val_drivers.

    * jllutils.stats York fitting procedure now does some input checking
to make sure that all input x, std_x, y, and std_y values are the same shape,
that all their unmasked values are finite, and (for std_x and std_y) > 0. This
should prevent issues that arose during testing where the fitting procedure
failed to converge because NaNs in the input caused the weights to fill with
NaNs.