Improved flexibility of York linear fit
Can now specify a convergence limit, as well as use the
`verbose` keyword to print information about how the fit
is iterating.

These changes were necessary because I ran into a case where
the residual was just above the convergence limit and oscillated
between positive and negative until the iteration limit was
reached. The `verbose` keyword helped figure this out and the
convergence limit option let me open up the convergence limit
so that it would actually realized it had converged.