Added functions for comparing arrays and netcdf files
In `miscutils`, I've added functions to compare regular and masked numpy arrays that should
take care of different types transparently. Additionally, the masked comparison should correctly
handle arrays with a scalar `np.False_` mask and an array of all `False` values.

In `subutils.ncdf`, I've added a function to compare data values in two netCDF files, and to
subset one or both before comparing, if needed. It seems that differently sized arrays cause
an error, rather than a false check, but that might be a good thing.